+++ src/pages/matching/components/OpponentInfo.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const OpponentInfo = () => {
+  return (
+    <div>
+      {/* Opponent information will be displayed here */}
+    </div>
+  );
+};
+
+export default OpponentInfo;
